fix(wallet-core): move thunk from actions to thunks to prevent circular esm
